why-use-mockbean/src/main/java/com/jojoldu/why/domain/book/BookService.java
package com.jojoldu.why.domain.book;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * Created by jojoldu@gmail.com on 2017. 10. 1.
 * Blog : http://jojoldu.tistory.com
 * Github : https://github.com/jojoldu
 */

@Service
public class BookService {

    @Autowired
    private BookRepository bookRepository;

//    public BookService(BookRepository bookRepository) {
//        this.bookRepository = bookRepository;
//    }

    public List<Book> findAll() {
        return bookRepository.findAll();
    }
}
